Finished basic actor

+++ Survivors Game AI/Assets/Scripts/ScriptableObjects/StatsSO.cs
@@ -6,15 +6,15 @@ using UnityEngine;
 public class StatsSO : ScriptableObject
 {
     [Header ("Main Stats")]
-    public float health = 0;
+    public int health = 0;
     public float movementSpeed = 0;
     [Header("Offensive Stats")]
-    public float damage = 0;
-    public float magicPower = 0;
+    public int damage = 0;
+    public int magicPower = 0;
     public float attackSpeed = 0;
-    public float critChance = 0;
-    public float critDamageMod = 2;
+    public int critChance = 0;
+    public int critDamageMod = 2;
     [Header("Defensive Stats")]
-    public float magicDefense = 0;
-    public float defense = 0;
+    public int magicDefense = 0;
+    public int defense = 0;
 }
